feat: enhance blog components and improve styling
- Updated the "GitOps at Home" post with improved formatting and expanded content.
- Refactored blog components to utilize a centralized code font constant for consistency.
- Introduced a new `EmptyPostsState` component to handle cases with no posts found.
- Simplified the `PostGrid` and `PostList` components by integrating the new empty state component.
- Added date validation in the blog compilation script to ensure proper post date formatting.
- Removed unused code and dependencies to streamline the blog library.
