+++ src/app/blog/components/listing/PostGrid.tsx
@@ -1,41 +1,20 @@
 "use client";
 
 import React from "react";
-import { Box, Typography, Grid, useTheme, alpha } from "@mui/material";
+import { Grid } from "@mui/material";
 import { motion } from "framer-motion";
 import { staggerContainer } from "@/app/lib/animations";
 import { Post } from "@/app/types/blog";
 import { BlogCard } from "./BlogCard";
+import { EmptyPostsState } from "./EmptyPostsState";
 
 interface PostGridProps {
   posts: Post[];
 }
 
 export function PostGrid({ posts }: PostGridProps) {
-  const theme = useTheme();
-
   if (posts.length === 0) {
-    return (
-      <Box
-        sx={{
-          width: "100%",
-          textAlign: "center",
-          py: 8,
-          px: 3,
-          backgroundColor: alpha(theme.palette.background.paper, 0.4),
-          borderRadius: "6px",
-          border: `1px dashed ${theme.palette.divider}`,
-          color: theme.palette.text.secondary,
-        }}
-      >
-        <Typography variant="h6" sx={{ fontWeight: 500 }}>
-          No posts found matching your criteria.
-        </Typography>
-        <Typography variant="body2" sx={{ mt: 1 }}>
-          Try selecting different tags or view all posts.
-        </Typography>
-      </Box>
-    );
+    return <EmptyPostsState />;
   }
 
   return (
