--- src/app/lib/code-font-constant.js
@@ -1,4 +1,0 @@
-/** @type {{ CODE_FONT_FAMILY: string }} */
-module.exports = {
-  CODE_FONT_FAMILY: 'var(--font-fira-code), "Geist Mono", monospace',
-};
